Add files via upload
The class diagram shows how certain classes have lists that are composed of other classes, and many of the classes are dependent on the main class to function. The sequence diagram shows how the system runs as the user is selecting options and entering information. It shows which classes send messages and information to other classes and the messages displayed to the user in the process.